rutabaga_gfx/ffi: Use type u8 for ring_idx
To be consistent with the protocol and other part of the code, change
type of ring_idx in rutabaga_gfx ffi to u8. This also fix an issue of
incorrect ring_idx in the case running QEMU with virtio-gpu-rutabaga
(venus enabled).

+++ rutabaga_gfx/ffi/src/include/rutabaga_gfx_ffi.h
@@ -186,7 +186,7 @@ struct rutabaga_fence {
     uint32_t flags;
     uint64_t fence_id;
     uint32_t ctx_id;
-    uint32_t ring_idx;
+    uint8_t ring_idx;
 };
 
 struct rutabaga_debug {
